Add database migration for subscription list email addresses
