+++ package.json
@@ -23,6 +23,9 @@
     "build-emails": "lerna run build-emails --scope server",
     "build-client": "lerna run build-client --scope client",
     "dev": "lerna run dev --scope server --scope client --stream --parallel",
+    "migrate-create": "migrate create --migrations-dir=packages/server/db/migrations",
+    "migrate-down": "migrate down --migrations-dir=packages/server/db/migrations",
+    "migrate-up": "migrate up --migrations-dir=packages/server/db/migrations",
     "start": "lerna run start --scope server --stream",
     "style": "eslint --quiet packages/server/**/*.js packages/client/**/*.js",
     "format": "prettier --write packages/server/**/*.js packages/client/**/*.js",
